refactor(cli): split Textual app into focused modules
